+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Breaking changes
 
   * GraphQL support has been updated to use `graphql_client` 0.14.
+  * The `http` crate has been updated to 1.x.
 
 # v0.1610.0